% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AddMeta.R
\name{AddMeta}
\alias{AddMeta}
\alias{.addMeta}
\alias{AddMeta,BRIC-method}
\title{Create addmeta function for adding meta information}
\usage{
AddMeta(object, ...)

.addMeta(object = NULL, meta.info = NULL)

\S4method{AddMeta}{BRIC}(object = NULL, meta.info = NULL)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{meta.info}{meta information table should be a data frame with rows representing cell and coloumn representing different group condition}
}
\value{
It will add meta informatoin to IRISFGM.
}
\description{
This function can import cell annotation information to IRIS-FGM object.
}
\examples{
\dontrun{AddMeta(object, meta.info = my.meta)}  
}
