% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DifferentialGene.R
\name{FindMarkers}
\alias{FindMarkers}
\alias{.findMarkers}
\alias{FindMarkers,BRIC-method}
\title{Find marker for idnetified cell clusters}
\usage{
FindMarkers(object, ...)

.findMarkers(object, SimpleResult = T, FDR = 0.05)

\S4method{FindMarkers}{BRIC}(object, SimpleResult = T, FDR = 0.05)
}
\arguments{
\item{object}{input IRIS-FGM object}

\item{SimpleResult}{marker gene only output log fold change (LFC), p-value, and adjusted p-value.}

\item{FDR}{a number to specify the threshold of FDR, default by 0.05}
}
\description{
This function has wrapped DEsingle package and perform single cell analysis based on LTMG signaling matrix.
}
\examples{
\dontrun{object <- FindMarker(object,SimpleResult = T, FDR = 0.05)}

}
